requirements.txt
blinker==1.7.0
click==8.1.7
coverage==7.4.0
Flask==3.0.0
Flask-Caching==2.1.0
Flask-SQLAlchemy==3.1.1
redis==5.0.1
greenlet==3.0.1
numpy==1.26.2
pandas==1.5.3
psycopg==3.1.13
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.8.0
Werkzeug==3.0.1
